Update Custom.js
Add functionality to "add" button. On click triggers updating of the original custom pizza in menu.js. The price is updated to totalCustom from context.js which stores the original price plus the price of the toppings added by the user.  This updated version is sent to the Checkout.js state management and the item is correctly reflected in the cart.