Update reducerCustom.js
Added additional functionality to ADD_RADIO_TOPPING. Initially, there was a persisting error with the radio buttons for toppings in the sauce category. This was corrected by hard coding the radio button functionality, ensuring only one sauce can be selected at a time and persisting the same logic after the pizza has been completed and the user wishes to create a new pizza.